商品出品機能の修正実装②
