+++ app/javascript/packs/application.js
@@ -1,12 +1,17 @@
-window.addEventListener('load', () => {
-  const priceInput = document.getElementById("item-price");
+// This file is automatically compiled by Webpack, along with any other files
+// present in this directory. You're encouraged to place your actual application logic in
+// a relevant structure within app/javascript and only use these pack files to reference
+// that code so it'll be compiled.
 
-   priceInput.addEventListener("input", () => {
-     const addTaxDom = document.getElementById("add-tax-price");
+require("@rails/ujs").start()
 
-     addTaxDom.innerHTML = Math.round(priceInput.value * 0.1 );
-     const addPofitDom = document.getElementById("profit");
+require("@rails/activestorage").start()
+require("channels")
 
-     addPofitDom.innerHTML = Math.round(priceInput.value - Math.round(priceInput.value * 0.1 ))
- })
-});
+
+// Uncomment to copy all static images under ../images to the output folder and reference
+// them with the image_pack_tag helper in views (e.g <%= image_pack_tag 'rails.png' %>)
+// or the `imagePath` JavaScript helper below.
+//
+// const images = require.context('../images', true)
+// const imagePath = (name) => images(name, true)
