+++ app/views/items/new.html.erb
@@ -87,6 +87,7 @@
     <%# 販売価格 %>
     <div class="sell-price">
       <div class="weight-bold-text question-text">
+      <%= javascript_pack_tag 'fee_calculation' %>
         <span>販売価格<br>(¥300〜9,999,999)</span>
         <a class="question" href="#">?</a>
       </div>
